+++ src/api/FotoGen.Infrastructure/Replicate/GetTrainedModelStatus/GetTrainedModelStatusMapper.cs
@@ -1,0 +1,11 @@
+using FotoGen.Domain.Entities.Models;
+
+namespace FotoGen.Infrastructure.Replicate.GetTrainedModelStatus;
+
+public class GetTrainedModelStatusMapper
+{
+    public static QueryModelTrainingStatus ToResponseDto(GetTrainedModelStatusResponse response)
+    {
+        return new QueryModelTrainingStatus(response.Id, response.Status, response.Model, response.Version);
+    }
+}
