+++ src/api/FotoGen.Infrastructure/Repositories/TrainedModelRepository.cs
@@ -1,9 +1,3 @@
-using System;
-using System.Collections.Generic;
-using System.Globalization;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
 using FotoGen.Domain.Entities;
 using FotoGen.Domain.Interfaces;
 using FotoGen.Domain.ValueObjects;
